# [fileconsumer] Generlize emit function