--- a/pkg/stanza/fileconsumer/emit/emit.go
+++ b/pkg/stanza/fileconsumer/emit/emit.go
@@ -1,0 +1,10 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package emit // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/emit"
+
+import (
+	"context"
+)
+
+type Callback func(ctx context.Context, token []byte, attrs map[string]any)