--- a/pkg/stanza/fileconsumer/reader.go
+++ b/pkg/stanza/fileconsumer/reader.go
@@ -12,6 +12,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/emit"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/fingerprint"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/scanner"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
@@ -19,9 +20,13 @@ import (
 )
 
 type readerConfig struct {
-	fingerprintSize int
-	maxLogSize      int
-	emit            EmitFunc
+	fingerprintSize         int
+	maxLogSize              int
+	emit                    emit.Callback
+	includeFileName         bool
+	includeFilePath         bool
+	includeFileNameResolved bool
+	includeFilePathResolved bool
 }
 
 // Reader manages a single file
@@ -33,13 +38,13 @@ type Reader struct {
 	lineSplitFunc bufio.SplitFunc
 	splitFunc     bufio.SplitFunc
 	encoding      helper.Encoding
-	processFunc   EmitFunc
+	processFunc   emit.Callback
 
 	Fingerprint    *fingerprint.Fingerprint
 	Offset         int64
 	generation     int
 	file           *os.File
-	FileAttributes *FileAttributes
+	FileAttributes map[string]any
 	eof            bool
 
 	HeaderFinalized bool
@@ -92,7 +97,7 @@ func (r *Reader) ReadToEnd(ctx context.Context) {
 		if err != nil {
 			r.Errorw("decode: %w", zap.Error(err))
 		} else {
-			r.processFunc(ctx, r.FileAttributes, token)
+			r.processFunc(ctx, token, r.FileAttributes)
 		}
 
 		if r.recreateScanner {
@@ -116,7 +121,7 @@ func (r *Reader) ReadToEnd(ctx context.Context) {
 // consumeHeaderLine checks if the given token is a line of the header, and consumes it if it is.
 // The return value dictates whether the given line was a header line or not.
 // If false is returned, the full header can be assumed to be read.
-func (r *Reader) consumeHeaderLine(ctx context.Context, _ *FileAttributes, token []byte) {
+func (r *Reader) consumeHeaderLine(ctx context.Context, token []byte, _ map[string]any) {
 	if !r.headerSettings.matchRegex.Match(token) {
 		// Finalize and cleanup the pipeline
 		r.HeaderFinalized = true
@@ -154,7 +159,7 @@ func (r *Reader) consumeHeaderLine(ctx context.Context, _ *FileAttributes, token
 
 	// Copy resultant attributes over current set of attributes (upsert)
 	for k, v := range ent.Attributes {
-		r.FileAttributes.HeaderAttributes[k] = v
+		r.FileAttributes[k] = v
 	}
 }