--- a/pkg/stanza/operator/input/file/config.go
+++ b/pkg/stanza/operator/input/file/config.go
@@ -43,25 +43,6 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		return nil, err
 	}
 
-	var preEmitOptions []preEmitOption
-
-	if fileconsumer.AllowHeaderMetadataParsing.IsEnabled() {
-		preEmitOptions = append(preEmitOptions, setHeaderMetadata)
-	}
-
-	if c.IncludeFileName {
-		preEmitOptions = append(preEmitOptions, setFileName)
-	}
-	if c.IncludeFilePath {
-		preEmitOptions = append(preEmitOptions, setFilePath)
-	}
-	if c.IncludeFileNameResolved {
-		preEmitOptions = append(preEmitOptions, setFileNameResolved)
-	}
-	if c.IncludeFilePathResolved {
-		preEmitOptions = append(preEmitOptions, setFilePathResolved)
-	}
-
 	var toBody toBodyFunc = func(token []byte) interface{} {
 		return string(token)
 	}
@@ -74,9 +55,8 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 	}
 
 	input := &Input{
-		InputOperator:  inputOperator,
-		toBody:         toBody,
-		preEmitOptions: preEmitOptions,
+		InputOperator: inputOperator,
+		toBody:        toBody,
 	}
 
 	input.fileConsumer, err = c.Config.Build(logger, input.emit)