… .chloggen/fileconsumer-emitfunc-generalize.yaml
… pkg/stanza/fileconsumer/file_test.go
… pkg/stanza/fileconsumer/reader_test.go
… pkg/stanza/fileconsumer/util_test.go
… pkg/stanza/operator/input/file/config_test.go